Rename database pkg
